# Cache more Altair functions