Pyspec optimization

#### Benchmarks

There is some improvement, although it's not very impressive.

##### pytest, single process
```sh
pytest --disable-bls eth2spec/test  --fork=altair
```
- before: 471 passed, 236 skipped, 1 warning in 413.84s (0:06:53)
- after:  471 passed, 236 skipped, 1 warning in 325.45s (0:05:25)

##### pytest, multiple processes
```sh
python -m pytest -n 8 --disable-bls eth2spec/test  --fork=altair
```
- before: 471 passed, 236 skipped, 9 warnings in 79.38s (0:01:19)
- after: 471 passed, 236 skipped, 9 warnings in 72.48s (0:01:12)

- TODO
    - [ ] probably decrease the LRU size